requirements.txt
Flask==2.1.2
python_bcrypt==0.3.2
shortuuid==1.0.11
sqlitedict==2.0.0
websockets==10.3
